Update and add test scnario

+++ classes.h
@@ -5,19 +5,19 @@
 
 typedef struct Unit{
     char isim[30];
-    int saldiri;
-    int savunma;
-    int saglik;
+    float saldiri;
+    float savunma;
+    float saglik;
     float kritik_sans;
 
-    int sayi;
+    float sayi;
 }Unit;
 
 
 typedef struct Hero{
     char isim[100];
     char bonus_turu[100];
-    int bonus_degeri;
+    float bonus_degeri;
     char etkilenen[100];
 }Hero;
 
@@ -25,7 +25,7 @@ typedef struct Hero{
 typedef struct Monester{
     char isim[100];
     char etki_turu[100];
-    int etki_degeri;
+    float etki_degeri;
     char etkilenen[100];
 }Monester;
 
@@ -35,7 +35,7 @@ typedef struct Monester{
 typedef struct Research{
     char isim[50];
     int seviye;
-    int oran;
+    float oran;
 
     char etkilenen[50];
 
@@ -51,9 +51,9 @@ typedef struct Takim{
     Monester canavarlar[11]; // max canavar :11
     int canavarSayisi;
     Research arastirma_seviyesi; // sadece bir tan olabilir
-    int totalHealth;
-    int totalAct;
-    int totalDefence;
+    float totalHealth;
+    float totalAct;
+    float totalDefence;
 }Takim;
